make remote agent tool calling class lightweight
